Fix warning messages 3

+++ cpp/brain/Actuator.h
@@ -36,7 +36,7 @@ namespace revolve
       virtual void update(double *output_vector,
                           double step) = 0;
 
-      virtual unsigned int outputs() const = 0;
+      virtual size_t outputs() const = 0;
     };
 
     typedef boost::shared_ptr<Actuator> ActuatorPtr;
