+++ cpp/test/test_CPGBrain.cpp
@@ -71,8 +71,8 @@ int main()
 
 TestCPGBrain::TestCPGBrain(std::string robot_name,
                            revolve::brain::EvaluatorPtr evaluator,
-                           unsigned int n_actuators,
-                           unsigned int n_sensors)
+                           size_t n_actuators,
+                           size_t n_sensors)
         : revolve::brain::CPGBrain(robot_name,
                                    evaluator,
                                    n_actuators,
